new file:   .gdb_history 	deleted:    a.out 	modified:   test.c

+++ test.c
@@ -3,6 +3,6 @@ int main() {
 	int x = 10, y = 20;
 	int z;
 	z = x * y + 3.0/2 * 4;
-	printf("%d\n", x * y + 3.0/2 * 4);
+	printf("%d\n", (int)(x * y + 3.0/2 * 4));
 	return 0;
 }
